Split RecordModifier into Edit and Save classes
